+++ packages/merkleize-metadata/src/main.ts
@@ -20,10 +20,11 @@ import {
 import { getAccessibleTypes } from "./get-accessible-types"
 import { getLookup } from "./get-lookup"
 import { getMetadata } from "./get-metadata"
-import { compactTypeRefs, mergeUint8, toBytes } from "./utils"
+import { compactTypeRefs, toBytes } from "./utils"
 import { decodeAndCollectKnownLeafs } from "./decode-and-collect"
 import { getProofData } from "./proof"
 import { getDynamicBuilder, getLookupFn } from "@polkadot-api/metadata-builders"
+import { mergeUint8 } from "@polkadot-api/utils"
 
 export interface MetadataMerkleizer {
   /**
@@ -190,7 +191,7 @@ export const merkleizeMetadata = (
 
     for (let i = hashTree.length - 2; i > 0; i -= 2)
       hashTree[(i - 1) / 2] = Blake3256(
-        mergeUint8([hashTree[i], hashTree[i + 1]]),
+        mergeUint8(hashTree[i], hashTree[i + 1]),
       )
 
     return hashTree
@@ -219,7 +220,7 @@ export const merkleizeMetadata = (
     const hashTree = getHashTree()
     const proofs = proofData.proofIdxs.map((idx) => hashTree[idx])
 
-    return mergeUint8([
+    return mergeUint8(
       compact.enc(proofData.leaves.length),
       ...proofData.leaves,
       compact.enc(proofData.leafIdxs.length),
@@ -228,7 +229,7 @@ export const merkleizeMetadata = (
       ...proofs,
       extrinsicMetadata.enc(extrinsic),
       extraInfo.enc(info),
-    ])
+    )
   }
 
   const getProofForExtrinsicPayload = (
@@ -250,7 +251,7 @@ export const merkleizeMetadata = (
     includedInSignedData: Uint8Array | HexString,
   ) => {
     const bytes = mergeUint8(
-      [callData, includedInExtrinsic, includedInSignedData].map(toBytes),
+      ...[callData, includedInExtrinsic, includedInSignedData].map(toBytes),
     )
     return getProofForExtrinsicPayload(bytes)
   }
@@ -259,22 +260,23 @@ export const merkleizeMetadata = (
     transaction: Uint8Array | HexString,
     txAdditionalSigned?: Uint8Array | HexString,
   ) => {
-    let [, { version, signed }, bytes] = extrinsicDec(transaction)
+    let [, { version, type }, bytes] = extrinsicDec(transaction)
 
     if (version !== extrinsic.version)
       throw new Error("Incorrect extrinsic version")
 
-    const typeRefs: TypeRef[] = signed
-      ? [
-          extrinsic.addressTy,
-          extrinsic.signatureTy,
-          ...extrinsic.signedExtensions.map((x) => x.includedInExtrinsic),
-          extrinsic.callTy,
-        ]
-      : [extrinsic.callTy]
+    const typeRefs: TypeRef[] =
+      type === "signed"
+        ? [
+            extrinsic.addressTy,
+            extrinsic.signatureTy,
+            ...extrinsic.signedExtensions.map((x) => x.includedInExtrinsic),
+            extrinsic.callTy,
+          ]
+        : [extrinsic.callTy]
 
     if (txAdditionalSigned) {
-      bytes = mergeUint8([bytes, toBytes(txAdditionalSigned)])
+      bytes = mergeUint8(bytes, toBytes(txAdditionalSigned))
       typeRefs.push(
         ...extrinsic.signedExtensions.map((x) => x.includedInSignedData),
       )